1869 retter feil i team tiltak sitt kafkaformat

+++ src/main/kotlin/no/nav/tiltakspenger/saksbehandling/tiltaksdeltakelse/infra/kafka/teamtiltak/AvtaleDto.kt
@@ -15,8 +15,8 @@ data class AvtaleDto(
     val avtaleStatus: AvtaleStatus,
     val startDato: LocalDate?,
     val sluttDato: LocalDate?,
-    val stillingprosent: Int?,
-    val antallDagerPerUke: Int?,
+    val stillingprosent: Double?,
+    val antallDagerPerUke: Double?,
     val feilregistrert: Boolean,
 ) {
     enum class AvtaleStatus {
